Use selectors instead of containers for charts logics
